moved out common functions
